Refactor visualizer to split API specific code from API agnostic code.
