+++ src/allocator/dedicated_block_allocator/visualizer.rs
@@ -1,0 +1,8 @@
+use super::DedicatedBlockAllocator;
+use crate::visualizer::SubAllocatorVisualizer;
+
+impl SubAllocatorVisualizer for DedicatedBlockAllocator {
+    fn draw_base_info(&self, ui: &imgui::Ui) {
+        ui.text("Dedicated Block");
+    }
+}
